+++ plugin-gradle/src/test/java/com/diffplug/gradle/spotless/KotlinExtensionTest.java
@@ -23,6 +23,8 @@ import java.time.YearMonth;
 import org.gradle.testkit.runner.BuildResult;
 import org.junit.Test;
 
+import com.diffplug.spotless.JreVersion;
+
 public class KotlinExtensionTest extends GradleIntegrationTest {
 	private static final String HEADER = "// License Header";
 	private static final String HEADER_WITH_YEAR = "// License Header $YEAR";
@@ -43,6 +45,28 @@ public class KotlinExtensionTest extends GradleIntegrationTest {
 		setFile("src/main/kotlin/basic.kt").toResource("kotlin/ktlint/basic.dirty");
 		gradleRunner().withArguments("spotlessApply").build();
 		assertFile("src/main/kotlin/basic.kt").sameAsResource("kotlin/ktlint/basic.clean");
+	}
+
+	@Test
+	public void integrationKtfmt() throws IOException {
+		if (JreVersion.thisVm() == JreVersion._8) {
+			// ktfmt's dependency, google-java-format 1.8 requires a minimum of JRE 11+.
+			return;
+		}
+		setFile("build.gradle").toLines(
+				"plugins {",
+				"    id 'nebula.kotlin' version '1.0.6'",
+				"    id 'com.diffplug.gradle.spotless'",
+				"}",
+				"repositories { mavenCentral() }",
+				"spotless {",
+				"    kotlin {",
+				"        ktfmt()",
+				"    }",
+				"}");
+		setFile("src/main/kotlin/basic.kt").toResource("kotlin/ktfmt/basic.dirty");
+		gradleRunner().withArguments("spotlessApply").build();
+		assertFile("src/main/kotlin/basic.kt").sameAsResource("kotlin/ktfmt/basic.clean");
 	}
 
 	@Test
@@ -83,6 +107,29 @@ public class KotlinExtensionTest extends GradleIntegrationTest {
 	}
 
 	@Test
+	public void testWithHeaderKtfmt() throws IOException {
+		if (JreVersion.thisVm() == JreVersion._8) {
+			// ktfmt's dependency, google-java-format 1.8 requires a minimum of JRE 11+.
+			return;
+		}
+		setFile("build.gradle").toLines(
+				"plugins {",
+				"    id 'nebula.kotlin' version '1.0.6'",
+				"    id 'com.diffplug.gradle.spotless'",
+				"}",
+				"repositories { mavenCentral() }",
+				"spotless {",
+				"    kotlin {",
+				"        licenseHeader('" + HEADER + "')",
+				"        ktfmt()",
+				"    }",
+				"}");
+		setFile("src/main/kotlin/test.kt").toResource("kotlin/licenseheader/KotlinCodeWithoutHeader.test");
+		gradleRunner().withArguments("spotlessApply").build();
+		assertFile("src/main/kotlin/test.kt").hasContent(HEADER + "\n" + getTestResource("kotlin/licenseheader/KotlinCodeWithoutHeaderKtfmt.test"));
+	}
+
+	@Test
 	public void testWithCustomHeaderSeparator() throws IOException {
 		setFile("build.gradle").toLines(
 				"plugins {",
@@ -102,6 +149,29 @@ public class KotlinExtensionTest extends GradleIntegrationTest {
 	}
 
 	@Test
+	public void testWithCustomHeaderSeparatorKtfmt() throws IOException {
+		if (JreVersion.thisVm() == JreVersion._8) {
+			// ktfmt's dependency, google-java-format 1.8 requires a minimum of JRE 11+.
+			return;
+		}
+		setFile("build.gradle").toLines(
+				"plugins {",
+				"    id 'nebula.kotlin' version '1.0.6'",
+				"    id 'com.diffplug.gradle.spotless'",
+				"}",
+				"repositories { mavenCentral() }",
+				"spotless {",
+				"    kotlin {",
+				"        licenseHeader ('" + HEADER + "', '@file')",
+				"        ktfmt()",
+				"    }",
+				"}");
+		setFile("src/main/kotlin/test.kt").toResource("kotlin/licenseheader/KotlinCodeWithoutHeader.test");
+		gradleRunner().withArguments("spotlessApply").build();
+		assertFile("src/main/kotlin/test.kt").hasContent(HEADER + "\n" + getTestResource("kotlin/licenseheader/KotlinCodeWithoutHeaderKtfmt.test"));
+	}
+
+	@Test
 	public void testWithNonStandardYearSeparator() throws IOException {
 		setFile("build.gradle").toLines(
 				"plugins {",
@@ -113,6 +183,36 @@ public class KotlinExtensionTest extends GradleIntegrationTest {
 				"    kotlin {",
 				"        licenseHeader('" + HEADER_WITH_YEAR + "').yearSeparator(', ')",
 				"        ktlint()",
+				"    }",
+				"}");
+
+		setFile("src/main/kotlin/test.kt").toResource("kotlin/licenseheader/KotlinCodeWithMultiYearHeader.test");
+		setFile("src/main/kotlin/test2.kt").toResource("kotlin/licenseheader/KotlinCodeWithMultiYearHeader2.test");
+		gradleRunner().withArguments("spotlessApply").build();
+		assertFile("src/main/kotlin/test.kt").matches(matcher -> {
+			matcher.startsWith("// License Header 2012, 2014");
+		});
+		assertFile("src/main/kotlin/test2.kt").matches(matcher -> {
+			matcher.startsWith(HEADER_WITH_YEAR.replace("$YEAR", String.valueOf(YearMonth.now().getYear())));
+		});
+	}
+
+	@Test
+	public void testWithNonStandardYearSeparatorKtfmt() throws IOException {
+		if (JreVersion.thisVm() == JreVersion._8) {
+			// ktfmt's dependency, google-java-format 1.8 requires a minimum of JRE 11+.
+			return;
+		}
+		setFile("build.gradle").toLines(
+				"plugins {",
+				"    id 'nebula.kotlin' version '1.0.6'",
+				"    id 'com.diffplug.gradle.spotless'",
+				"}",
+				"repositories { mavenCentral() }",
+				"spotless {",
+				"    kotlin {",
+				"        licenseHeader('" + HEADER_WITH_YEAR + "').yearSeparator(', ')",
+				"        ktfmt()",
 				"    }",
 				"}");