+++ CHANGES.md
@@ -10,9 +10,12 @@ This document is intended for Spotless developers.
 We adhere to the [keepachangelog](https://keepachangelog.com/en/1.0.0/) format (starting after version `1.27.0`).
 
 ## [Unreleased]
+
+## [1.29.0] - 2020-05-05
 ### Added
 * Support for google-java-format 1.8 (including test infrastructure for Java 11). ([#562](https://github.com/diffplug/spotless/issues/562))
 * Improved PaddedCell such that it is as performant as non-padded cell - no reason not to have it always enabled.  Deprecated all of `PaddedCellBulk`. ([#561](https://github.com/diffplug/spotless/pull/561))
+* Support for ktfmt 0.13 ([#569](https://github.com/diffplug/spotless/pull/569))
 ### Changed
 * Updated a bunch of dependencies, most notably: ([#564](https://github.com/diffplug/spotless/pull/564))
   * jgit `5.5.0.201909110433-r` -> `5.7.0.202003110725-r`